files_to_server/requirements.txt
# Core ML stack
torch>=2.1.0
transformers>=4.43.0
trl>=0.9.6
accelerate>=0.33.0
peft>=0.11.1
bitsandbytes>=0.43.0
huggingface_hub
# Data handling
pandas>=2.0.0
datasets>=2.20.0

# Tokenization utilities
sentencepiece>=0.1.99

# Optional but recommended for HF models
safetensors>=0.4.3

# basic package
python-dotenv